Add handlers to delegate validators
CerberusValidator looks for the method name on the service component to
get the schema validators.
Delegate this to a Component, so we can customize the way the validator
for a service is found.

E.g.:

* for each service component, we have 2 additional components: a
  validator for the input schema and another for the output schema
* when looking for the validator, we have to search for the
  corresponding input/output validation component

As implemented in: https://github.com/OCA/wms/pull/137

By delegating this lookup to a base "cerberus.validator" component, we
can override the default behavior for a services collection, by
inheriting the Component.

By default, the behavior does not change.