Add direction in methods to get validator schema
For input (request) / output (response) validation schemas.
This direction is passed to the component BaseRestCerberusValidator,
which can be customized to use another strategy to retrieve the
validator (such as delegating to dedicated components).

The direction is not used in the default implementation as the
validation schema is get from the passed method name on the service.

An implementation using it could be: when the direction is "input", look
for a method name on a component named <service>.input and when the
direction is "output", look for a component named <service>.output.