[FIX] datamodel: Avoid maximum recursion
Into the datamodel constructor, we 'patch' the datamodel constructor to ensure the environment is propagated when a instance is created. Before this change, each time the factory was called, the init method was patched again. After some times, the maximum recusion limit was exceed. With this change we ensure that the patch is applied only once.