[FIX] base_rest_datamodel: remove incorrect validation of responses
In marshmallow, `dump` (serialization) and `load` (deserialization) operations using the same schema are not symmetrical in general. That is because fields can be declared as `dump_only` or `load_only`. That means that trying to `load` data you just `dumped` can result in errors, and vice-versa.

And that is exactly what is (was) wrong with this validation: the `validate` method on the `ModelClass` uses `load` to validate the data. That means that if you declared some fields to be `dump_only`, this validation necessarily crash.

Moreover, it seems that this validation is not only incorrect but also useless (if it worked as intended), since it bases its verification on the very same schema as the one that was just used used to serialize the data.

[FIX] base_rest_datamodel: change response validation to use a schema with `unknown=EXCLUDE`

This modification prevents fields that are declared as `dump_only` to mess with the validation (`dump_only` fields are seen as "unknown" during validation, since marshmallow validation is designed to be used during loading only).

[FIX] base_rest_datamodel: change response validation to use a schema with `unknown=EXCLUDE`

Nested model required extra care because the `unknown=EXCLUDE` is not passed down to the nested schemas when validating data, and this behavior could not easily be overridden. I ended up removing `dump_only` fields from value during `NestedModel` deserialization.

[ADD] base_rest_demo: add test for new api

[ADD] base_rest_demo: test
[FIX] datamodel: allow to instantiate datamodel with dump_only fields as well
[IMP] datamodel: improved 'validate' method, propagating 'unknown' value to nested schemas

[IMP] base_rest_demo: rename test class