[IMP] base_rest_model: allow to receive `partial` datamodels as params
In particular, this feature allows one to use datamodels to make partial updates through `PUT` requests. In this case, the endpoint declaration would look like this:

```
    @restapi.method(
        [(["/"], "PUT")],
        input_param=Datamodel("partner.info", partial=True),
        output_param=Datamodel("partner.info"),
    )
```

If `partial` is not set (or without this feature), the params sent by the client in the request must contain values for all the fields declared in the datamodel. Note that it uses an existing functionality of `marshmallow_objects`, as a call to `dump` of a `Datamodel` declared as `partial` would contain only the subset of fields provided in the params.

[FIX] base_rest_model: linting

[FIX] base_rest_model: linting

[FIX] base_rest_model: remove useless comment