[MIG] model_serializer: Migration to 15.0
GenericMinimalSerializer had to be adapted since it made use of the "id" and "display_name" field definitions on the "base" model. But since 15.0 those magic fields are not defined on abstract models anymore. They are added dynamically when the model class is created through the metaclass.

In order to work around the issue the simplest solution was to explicitly define the "id" and "display_name" marshmallow fields in GenericMinimalSerializer.