[IMP] fastapi: Improves UX for the endpoint registry synchronisation
A ribbon and a button are dislayed on the form view when a record needs to be synchronized.
In the tree view the records to synchronize are displayed with the text decoration defined for warning. A button is also displayed on the line if sync is required and an server actions allows you to sync selected records in one click